# Speed up feature extraction using batch level parallelisation with da…